+++ frontends/admin/react/components/molecules/index.ts
@@ -1,0 +1,9 @@
+export { Content } from "./content";
+export { ContentHeader } from "./content-header";
+export { DataTable } from "./data-table";
+export { ErrorBox } from "./error-box";
+export { Form } from "./form";
+export { ProtectedRoute } from "./protected-route";
+export { SideBar } from "./side-bar";
+export { StatCard } from "./stat-card";
+export { TopBar } from "./top-bar";