+++ frontends/admin/react/contexts/api.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+import { ColonyPrintAPI } from "../api/colony-print";
+
+export const APIContext = createContext<ColonyPrintAPI>(
+    new ColonyPrintAPI()
+);