+++ frontends/admin/react/contexts/data.ts
@@ -1,0 +1,11 @@
+import { createContext } from "react";
+
+export interface DataContextType {
+    username: string | null;
+    setUsername: (username: string | null) => void;
+}
+
+export const DataContext = createContext<DataContextType>({
+    username: null,
+    setUsername: () => {}
+});