+++ frontends/admin/react/hooks/use-api.ts
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+
+import { APIContext } from "../contexts";
+import { ColonyPrintAPI } from "../api/colony-print";
+
+export const useAPI = (): ColonyPrintAPI => {
+    return useContext(APIContext);
+};